refactor: renaming and moving provider to same file as context
